Parse block expression, fix spaghetti code
Previously, a call to parse_block_expression had to handle the fact that
the expression was surrouned in braces. Any block expression is
surrouned in braces, so it makes more sense to move this check to
parse_block_expression, such that the current token is the left brace
when the call starts and when it ends the current token is the first
token after the right brace.