Add (broken) test for recursion
Currently doesn't work because at the time of creating a function, the
function itself isn't in the environment.

Can be solved by only creating an environment at call time, but then
closures don't work anymore.

Both can work at the same time when following the book's nested
environment example, but that's a mess with ownership in Rust, that is,
shared mutable state is evil.

Hacky solution that does work how you want it to is to inject self into
environment at evaluation time, but a) that creates a ton of memory
garbage because environment has to be cloned and b) currently we don't
know the function's name as functions are just expressions and there's
no clean way to associate a function's name with its value.